still adding code to schedule service

+++ src/logistics/facilityservice/Facility.java
@@ -6,6 +6,6 @@ package logistics.facilityservice;
 public interface Facility
 {
     String getFacilityName();
-    Double getFacilityRate();
+    Integer getFacilityRate();
     Double getFacilityCost();
 }
